requirements.txt
pydantic==2.10.6
torch==2.6.0
pillow==11.1.0
torchvision==0.21.0
chromadb==0.6.3
requests==2.32.3
streamlit==1.43.2